requirements-dev.txt
# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==4.2.9  # AWS service mocking
boto3-stubs[ses,dynamodb]==1.34.0

# Linting and formatting
pylint==3.0.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Type hints
types-boto3==1.0.2

# Security scanning
bandit[toml]==1.7.5
safety==2.3.5

# Integration testing
requests==2.31.0
